requirements.txt
asgiref==3.3.1
Django==3.1.6
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
pandas
python-dateutil==2.8.1
pytz==2021.1
rest-pandas==1.1.0
six==1.15.0
sqlparse==0.4.1
unicodecsv==0.14.1
